- Added debugger support to duk.

+++ duk/options.h
@@ -6,11 +6,11 @@
 
 struct Options final
 {
-  static constexpr unsigned int defaultDebugPort = 37851;
+  static constexpr unsigned short defaultDebugPort = 37851;
 
   std::string input;
   bool debug;
-  unsigned int debugPort;
+  unsigned short debugPort;
 };
